Update App.js
Added the dynamic styles for you. i wasn't sure you were in the virtual class. now you should go to your css file and add a .red style pointer, set the color to red, add another pointer .bold and set the font weight to bold and a last pointer .wahala set the text decoration to underline and see the result